Restructure; delint.
